+++ src/templates/common/mediaImage/index.test.tsx
