+++ src/templates/common/mediaImage/index.tsx
