+++ src/templates/common/mediaImage/media.stories.tsx
