+++ src/templates/common/mediaImage/mockMedia.json
